Finishing Friday assignment in Week14 folder

+++ Week14/JPAAssignmentThursday/src/main/java/com/skillstorm/controllers/CustomerController.java
@@ -29,7 +29,7 @@ public class CustomerController {
     }
     
     @GetMapping("/{id}")
-    public Customer findById(@PathVariable int id) {
+    public Customer findById(@PathVariable Long id) {
         return service.findById(id);
     }
