+++ Week14/JPAAssignmentThursday/src/main/java/com/skillstorm/controllers/GroceryItemController.java
@@ -28,7 +28,7 @@ public class GroceryItemController {
     }
     
     @GetMapping("/{id}")
-    public GroceryItem findById(@PathVariable int id) {
+    public GroceryItem findById(@PathVariable Long id) {
         return service.findById(id);
     }
