+++ Week14/JPAAssignmentThursday/src/main/java/com/skillstorm/controllers/OrderInfoController.java
@@ -28,7 +28,7 @@ public class OrderInfoController {
     }
     
     @GetMapping("/{id}")
-    public OrderInfo findById(@PathVariable int id) {
+    public OrderInfo findById(@PathVariable Long id) {
         return service.findById(id);
     }
